+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Properly display unknown editor code block languages as `Unknown`.
 - Fix missing editor highlight when selecting a separator.
 - Fix content update issue on slow connections.
+- Improve note editor focus restoration mechanism.
 
 ## v0.1.5